take into account review comments
