+++ rs_dpr_service/conversion_processor.py
@@ -1,0 +1,164 @@
+# Copyright 2024 CS Group
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+""" Conversion Processor for converting a legacy product (safe format) into new Zarr format """
+import uuid
+import fsspec
+from dask.distributed import Client
+from starlette.requests import Request
+from pygeoapi.util import JobStatus
+from pygeoapi.process.manager.postgresql import (
+    PostgreSQLManager,  # pylint: disable=C0302
+)
+from rs_dpr_service.processors import GeneralProcessor
+from rs_dpr_service.call_dask import convert_safe_to_zarr
+
+
+class ConversionProcessor(GeneralProcessor):
+    """Runs an legacy product (safe format) conversion into new zarr format as a Dask job via subprocess."""
+
+    def __init__(
+        self,
+        credentials: Request,
+        db_process_manager: PostgreSQLManager,
+    ):
+        """
+        Initialize Conversion Processor
+        """
+        super().__init__(credentials, db_process_manager, "ConversionProcessor")
+
+    def _check_safe_s3_config(self, data: dict):
+        """Validate the S3 bucket credentials for the legacy product."""
+        try:
+            safe_s3_cfg = data.get("safe_s3_config", {})
+            safe_s3_config = {
+                "key": safe_s3_cfg["key"],
+                "secret": safe_s3_cfg["secret"],
+                "client_kwargs": {
+                    "endpoint_url": safe_s3_cfg["client_kwargs"]["endpoint_url"],
+                    "region_name": safe_s3_cfg["client_kwargs"]["region_name"],
+                },
+            }
+        except (KeyError, TypeError) as e:
+            raise ValueError(f"Missing safe S3 config parameter: {e}") from e
+
+        try:
+            fs = fsspec.filesystem("s3", **safe_s3_config)
+            fs.ls("/")
+            return fs
+        except Exception as e:
+            raise ConnectionError(f"Failed to connect to safe S3: {e}") from e
+
+    def _check_zarr_s3_config(self, data: dict):
+        """Validate the S3 bucket credentials for the output path."""
+        try:
+            zarr_s3_cfg = data.get("zarr_s3_config", {})
+            zarr_s3_config = {
+                "key": zarr_s3_cfg["key"],
+                "secret": zarr_s3_cfg["secret"],
+                "client_kwargs": {
+                    "endpoint_url": zarr_s3_cfg["client_kwargs"]["endpoint_url"],
+                    "region_name": zarr_s3_cfg["client_kwargs"]["region_name"],
+                },
+            }
+        except (KeyError, TypeError) as e:
+            raise ValueError(f"Missing zarr S3 config parameter: {e}") from e
+
+        try:
+            fs = fsspec.filesystem("s3", **zarr_s3_config)
+            fs.ls("/")
+            return fs
+        except Exception as e:
+            raise ConnectionError(f"Failed to connect to zarr S3: {e}") from e
+
+    def _check_input_output_uris(self, safe_fs, zarr_fs, data: dict):
+        """Check that input legacy product exists and output bucket path exists."""
+
+        safe_uri = data.get("input_safe_path", "")
+        out_dir = data.get("output_zarr_dir_path", "").rstrip("/")
+        if not safe_uri.startswith("s3://"):
+            raise ValueError(f"Invalid input_safe_path format (must start with 's3://'): {safe_uri}")
+        if not out_dir.startswith("s3://"):
+            raise ValueError(f"Invalid output_zarr_dir_path format (must start with 's3://'): {out_dir}")
+
+        path = safe_uri.replace("s3://", "")
+        if not safe_fs.exists(path):
+            raise FileNotFoundError(f"Input SAFE path does not exist: {safe_uri}")
+
+        bucket = out_dir.replace("s3://", "").split("/", 1)[0]
+        if not zarr_fs.exists(bucket):
+            raise FileNotFoundError(f"Output S3 bucket does not exist: {out_dir}")
+
+    def _check_write_permission(self, fs, out_dir: str):
+        """Check write permission on the output bucket."""
+        bucket = out_dir.replace("s3://", "").split("/", 1)[0]
+        test_key = f"{bucket}/.perm_test_{uuid.uuid4().hex}"
+        try:
+            with fs.open(test_key, "wb"):
+                pass
+            fs.rm(test_key, recursive=False)
+        except Exception as e:
+            if "AccessDenied" in str(e) or "UnauthorizedOperation" in str(e):
+                raise PermissionError(f"No write permission on bucket: {out_dir}") from e
+            raise RuntimeError(f"Error checking write permissions: {e}") from e
+
+    async def execute(
+        self,
+        data: dict,
+        outputs=None,
+    ) -> tuple[str, dict]:
+        """
+        Asynchronously execute the conversion process.
+        """
+        try:
+            safe_fs = self._check_safe_s3_config(data)
+            zarr_fs = self._check_zarr_s3_config(data)
+            self._check_input_output_uris(safe_fs, zarr_fs, data)
+            self._check_write_permission(zarr_fs, data["output_zarr_dir_path"])
+        except Exception as e:  # pylint: disable=broad-exception-caught
+            msg = str(e)
+            self.logger.error(f"Conversion failed: {msg}")
+            self.log_job_execution(JobStatus.failed, None, msg)
+            return self._get_execute_result()
+
+        return await super().execute(data, outputs)
+
+    def manage_dask_tasks(self, client: Client, dpr_payload: dict):
+        """
+        Schedule SAFE to Zarr conversion on the Dask cluster using a nested subprocess task.
+        """
+        self.log_job_execution(JobStatus.running, 5, "Preparing conversion")
+        try:
+            safe_uri = dpr_payload.get("input_safe_path")
+            out_dir = dpr_payload.get("output_zarr_dir_path", "").rstrip("/")
+            basename = str(safe_uri).rsplit("/", 1)[-1].split(".", 1)[0]
+            zarr_uri = f"{out_dir}/{basename}.zarr"
+
+            cfg = {
+                "safe_uri": safe_uri,
+                "zarr_uri": zarr_uri,
+                "safe_s3_config": dpr_payload.get("safe_s3_config", {}),
+                "zarr_s3_config": dpr_payload.get("zarr_s3_config", {}),
+            }
+
+            future = client.submit(convert_safe_to_zarr, cfg)
+            self.log_job_execution(JobStatus.running, 50, "Conversion job submitted to cluster")
+
+            res = future.result()
+            self.log_job_execution(JobStatus.successful, 100, res)
+        except Exception as e:  # pylint: disable=broad-exception-caught
+            self.logger.error(f"Conversion failed: {e}")
+            self.log_job_execution(JobStatus.failed, None, f"Conversion failed: {e}")
+        finally:
+            client.close()
