+++ rs_dpr_service/main.py
@@ -39,16 +39,24 @@ from starlette.status import (  # pylint: disable=C0411
 )
 
 from rs_dpr_service import opentelemetry
+from rs_dpr_service.conversion_processor import ConversionProcessor
 from rs_dpr_service.jobs_table import Base
 from rs_dpr_service.openapi_validation import (
     validate_request,
     validate_response,
 )
-from rs_dpr_service.processors import env_bool, processors
+from rs_dpr_service.processors import S1L0Processor, S3L0Processor, env_bool
 
 # flake8: noqa: F401
 # DON'T REMOVE (needed for SQLAlchemy)
 from . import jobs_table  # pylint: disable=unused-import
+
+# Register all the processors
+processors = {
+    "S1L0_processor": S1L0Processor,
+    "S3L0_processor": S3L0Processor,
+    "Conversion_Processor": ConversionProcessor,
+}
 
 # Initialize a FastAPI application
 app = FastAPI(title="rs-dpr-service", root_path="", debug=True)
